% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeV2.R
\name{CalculateDEScore}
\alias{CalculateDEScore}
\title{Calculates the DE score for a pair of clusters}
\usage{
CalculateDEScore(seurat.object, cluster1, cluster2, pct.1, min.log2.fc)
}
\arguments{
\item{seurat.object}{a normalized, integrated Seurat object}

\item{cluster1}{the first cluster in the pair of clusters}

\item{cluster2}{the second cluster in the pair of clusters}

\item{pct.1}{the minimum gene expression fraction required for genes used to calculate the DE score}

\item{min.log2.fc}{the minimum log2 fold change required for genes used to calculate the DE score}
}
\value{
the DE score between the pair of clusters
}
\description{
Calculates the DE score for a pair of clusters
}
