% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeV2.R
\name{FindCentroids}
\alias{FindCentroids}
\title{Calculates centroids for each cluster in dim.reduction space}
\usage{
FindCentroids(seurat.object, n.dims, dim.reduction)
}
\arguments{
\item{seurat.object}{a normalized, integrated Seurat object}

\item{n.dims}{number of dimensions used to calculate centroids}

\item{dim.reduction}{the dimensional reduction used to calculate centroids}
}
\value{
centroids for each cluster in dim.reduction space
}
\description{
Calculates centroids for each cluster in dim.reduction space
}
