% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeV2.R
\name{IterativeClustering}
\alias{IterativeClustering}
\title{Performs iterative Leiden clustering on a Seurat object}
\usage{
IterativeClustering(
  seurat.object,
  max.iterations = 20,
  min.cluster.size = 10,
  min.de.score = 150,
  pct.1 = 0.5,
  min.log2.fc = 2,
  n.dims = 30,
  dim.reduction = "pca"
)
}
\arguments{
\item{seurat.object}{a normalized, integrated Seurat object}

\item{max.iterations}{the maximum number of iterations to perform. Default 20.}

\item{min.cluster.size}{the minimum number of cells in a cluster. Default 10.}

\item{min.de.score}{the minimum DE score required to consider clusters biologically-distinct. Default 150.}

\item{pct.1}{the minimum gene expression fraction required for genes used to calculate the DE score. Default 0.5.}

\item{min.log2.fc}{the minimum log2 fold change required for genes used to calculate the DE score. Default 2.}

\item{n.dims}{number of dimensions used to create the sNN graph and calculate centroids. Default 30.}

\item{dim.reduction}{the dimensional reduction used to create the sNN graph and calculate centroids. Default pca.}
}
\value{
a Seurat object with iterative clustering results in the seurat_clusters metadata value
}
\description{
Performs iterative Leiden clustering on a Seurat object
}
