% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeV2.R
\name{RunClusteringIteration}
\alias{RunClusteringIteration}
\title{Runs a single Leiden clustering iteration}
\usage{
RunClusteringIteration(
  seurat.object,
  min.cluster.size,
  min.de.score,
  pct.1,
  min.log2.fc,
  n.dims,
  dim.reduction
)
}
\arguments{
\item{seurat.object}{a normalized, integrated Seurat object}

\item{min.cluster.size}{the minimum number of cells in a cluster}

\item{min.de.score}{the minimum DE score required to consider clusters biologically-distinct}

\item{pct.1}{the minimum gene expression fraction required for genes used to calculate the DE score}

\item{min.log2.fc}{the minimum log2 fold change required for genes used to calculate the DE score}

\item{n.dims}{number of dimensions used to create the sNN graph and calculate centroids}

\item{dim.reduction}{the dimensional reduction used to create the sNN graph and calculate centroids}
}
\value{
a Seurat object with the results of the iteration in the seurat_clusters metadata value
}
\description{
Runs a single Leiden clustering iteration
}
